Restructure Redux
