+++ webapp/src/action-types.js
@@ -1,0 +1,5 @@
+export const GET_COURSES_REQUEST = 'GET_COURSES_REQUEST';
+export const GET_COURSES_SUCCESS = 'GET_COURSES_SUCCESS';
+export const GET_COURSES_ERROR = 'GET_COURSES_ERROR';
+
+export const TOGGLE_THEME = 'TOGGLE_THEME';
