--- webapp/src/store/actions/user-settings.js
@@ -1,3 +1,0 @@
-export const TOGGLE_THEME = {
-  type: 'TOGGLE_THEME'
-};
