--- webapp/src/store/reducers/index.js
@@ -1,8 +1,0 @@
-import {combineReducers} from 'redux';
-import courses from './courses';
-import userSettings from './user-settings';
-
-export default combineReducers({
-  courses,
-  userSettings
-});
